apify/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
PyPDF2==3.0.1
apify-client>=1.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
